chore: migrate sveltekit to v2 for legend
