+++ packages/legend/package.json
@@ -37,15 +37,16 @@
 		"@arkweid/lefthook": "^0.7.7",
 		"@creativebulma/bulma-tooltip": "^1.2.0",
 		"@geoffcox/svelte-splitter": "^1.0.1",
-		"@sveltejs/adapter-auto": "^2.1.1",
-		"@sveltejs/kit": "^1.30.4",
+		"@sveltejs/adapter-auto": "^3.1.1",
+		"@sveltejs/kit": "^2.5.4",
 		"@sveltejs/package": "^2.3.0",
+		"@sveltejs/vite-plugin-svelte": "^3.0.2",
 		"@types/chroma-js": "^2.4.4",
 		"@types/js-yaml": "^4.0.9",
 		"@types/lodash-es": "^4.17.12",
 		"@types/svelte-range-slider-pips": "^2.0.4",
-		"@typescript-eslint/eslint-plugin": "^6.21.0",
-		"@typescript-eslint/parser": "^6.21.0",
+		"@typescript-eslint/eslint-plugin": "^7.2.0",
+		"@typescript-eslint/parser": "^7.2.0",
 		"@watergis/svelte-maplibre-menu": "workspace:^",
 		"@watergis/svelte-maplibre-style-switcher": "workspace:^",
 		"bulma": "^0.9.4",
@@ -53,7 +54,7 @@
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.35.1",
-		"pmtiles": "^2.11.0",
+		"pmtiles": "^3.0.5",
 		"prettier": "^3.2.5",
 		"prettier-plugin-svelte": "^3.2.2",
 		"sass": "^1.72.0",
@@ -62,7 +63,7 @@
 		"svelte-preprocess": "^5.1.3",
 		"tslib": "^2.6.2",
 		"typescript": "^5.4.2",
-		"vite": "^4.5.2"
+		"vite": "^5.1.6"
 	},
 	"type": "module",
 	"dependencies": {
@@ -72,9 +73,9 @@
 		"detect-touch-device": "^1.1.6",
 		"js-yaml": "^4.1.0",
 		"lodash-es": "^4.17.21",
-		"maplibre-gl": "^3.6.2",
+		"maplibre-gl": "^4.1.0",
 		"svelte-awesome-color-picker": "^2.4.8",
-		"svelte-fa": "^3.0.4",
+		"svelte-fa": "^4.0.2",
 		"svelte-popperjs": "^1.3.2",
 		"svelte-range-slider-pips": "^2.3.1",
 		"svelte-tippy": "^1.3.2",
